* 登录失败锁定检查逻辑优化

+++ diboot-iam-starter/src/main/java/com/diboot/iam/config/Cons.java
@@ -134,5 +134,5 @@ public class Cons extends com.diboot.core.config.Cons {
     /**
      * 登录最大失败次数
      */
-    public static final long LOGIN_MAX_ATTEMPTS = 5;
+    public static final int LOGIN_MAX_ATTEMPTS = 5;
 }
